Fixed hero button and links with Auth

+++ app/Http/Controllers/ListingController.php
@@ -59,6 +59,12 @@ class ListingController extends Controller
 
     //update listings
     public function update(Request $request, Listing $listing){
+        
+        //Make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort('403', 'Unauthorized access');
+        }
+        
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -81,8 +87,19 @@ class ListingController extends Controller
     }
 
     //Delete listing
+
     public function destroy(Listing $listing){
+        //Make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort('403', 'Unauthorized access');
+        }
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted Successfully!');
+    }
+
+
+    //Manage Listings
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
